Moved ThrottleSettings to model and ThrottleRequests to util
